+++ test/smoke/paralleltestfixture_test.go
@@ -27,6 +27,7 @@ type (
 	}
 
 	ParallelTestFixture struct {
+		T                  *testing.T
 		NextAddr           func() string
 		testNames          map[string]struct{}
 		testNamesMu        sync.RWMutex
@@ -37,20 +38,21 @@ type (
 		testNamesFailed    map[string]struct{}
 		testNamesFailedMu  sync.Mutex
 	}
+
+	ParallelTestFixtureSet struct {
+		NextAddr func() string
+		mu       sync.Mutex
+		fixtures map[string]*ParallelTestFixture
+	}
 )
 
-func resetSingularity(t *testing.T) {
-	envDesc := getEnvDesc(t)
-	singularity := NewSingularity(envDesc.SingularityURL())
-	singularity.Reset(t)
-}
-
-func newParallelTestFixture(t *testing.T, opts PTFOpts) *ParallelTestFixture {
-	t.Helper()
-	resetSingularity(t)
-	stopPIDs(t)
+func newParallelTestFixtureSet(opts PTFOpts) *ParallelTestFixtureSet {
+	resetSingularity()
+	if err := stopPIDs(); err != nil {
+		panic(err)
+	}
 	numFreeAddrs := opts.NumFreeAddrs
-	freeAddrs := freePortAddrs(t, "127.0.0.1", numFreeAddrs, 6601, 9000)
+	freeAddrs := freePortAddrs("127.0.0.1", numFreeAddrs, 6601, 9000)
 	var nextAddrIndex int64
 	nextAddr := func() string {
 		i := atomic.AddInt64(&nextAddrIndex, 1)
@@ -59,13 +61,33 @@ func newParallelTestFixture(t *testing.T, opts PTFOpts) *ParallelTestFixture {
 		}
 		return freeAddrs[i]
 	}
-	return &ParallelTestFixture{
-		NextAddr:         nextAddr,
+	return &ParallelTestFixtureSet{
+		NextAddr: nextAddr,
+		fixtures: map[string]*ParallelTestFixture{},
+	}
+}
+
+func resetSingularity() {
+	envDesc := getEnvDesc()
+	singularity := NewSingularity(envDesc.SingularityURL())
+	singularity.Reset()
+}
+
+func (pfs *ParallelTestFixtureSet) newParallelTestFixture(t *testing.T) *ParallelTestFixture {
+	t.Helper()
+	t.Parallel()
+	pf := &ParallelTestFixture{
+		T:                t,
+		NextAddr:         pfs.NextAddr,
 		testNames:        map[string]struct{}{},
 		testNamesPassed:  map[string]struct{}{},
 		testNamesSkipped: map[string]struct{}{},
 		testNamesFailed:  map[string]struct{}{},
 	}
+	pfs.mu.Lock()
+	defer pfs.mu.Unlock()
+	pfs.fixtures[t.Name()] = pf
+	return pf
 }
 
 func (fcfg fixtureConfig) Desc() string {
@@ -84,6 +106,29 @@ func (pf *ParallelTestFixture) recordTestStarted(t *testing.T) {
 		t.Fatalf("duplicate test name: %q", name)
 	}
 	pf.testNames[name] = struct{}{}
+}
+
+// PTest is a test to run in parallel.
+type PTest struct {
+	Name string
+	Test func(*testing.T, *TestFixture)
+}
+
+func (pf *ParallelTestFixture) RunMatrix(m []fixtureConfig, tests ...PTest) {
+	for _, c := range m {
+		pf.T.Run(c.Desc(), func(t *testing.T) {
+			c := c
+			t.Parallel()
+			for _, pt := range tests {
+				pt := pt
+				t.Run(pt.Name, func(t *testing.T) {
+					f := pf.NewIsolatedFixture(t, c)
+					defer f.ReportStatus(t)
+					pt.Test(t, f)
+				})
+			}
+		})
+	}
 }
 
 func (pf *ParallelTestFixture) recordTestStatus(t *testing.T) {
@@ -115,17 +160,33 @@ func (pf *ParallelTestFixture) recordTestStatus(t *testing.T) {
 	}
 }
 
-func (pf *ParallelTestFixture) PrintSummary(t *testing.T) {
-	total := len(pf.testNames)
-	passed := len(pf.testNamesPassed)
-	skipped := len(pf.testNamesSkipped)
-	failed := len(pf.testNamesFailed)
+func (pfs *ParallelTestFixtureSet) PrintSummary() {
+	var total, passed, skipped, failed, missing int
+	for _, pf := range pfs.fixtures {
+		t, p, s, f, m := pf.PrintSummary()
+		total += t
+		passed += p
+		skipped += s
+		failed += f
+		missing += m
+	}
+	summary := fmt.Sprintf("Summary: %d failed; %d skipped; %d passed; %d missing (total %d)", failed, skipped, passed, missing, total)
+	fmt.Fprintln(os.Stdout, summary)
+}
 
-	summary := fmt.Sprintf("Test summary: %d failed; %d skipped; %d passed (total %d)", failed, skipped, passed, total)
+func (pf *ParallelTestFixture) PrintSummary() (total, passed, skipped, failed, missing int) {
+	t := pf.T
+	t.Helper()
+	total = len(pf.testNames)
+	passed = len(pf.testNamesPassed)
+	skipped = len(pf.testNamesSkipped)
+	failed = len(pf.testNamesFailed)
+
+	summary := fmt.Sprintf("%s summary: %d failed; %d skipped; %d passed (total %d)", t.Name(), failed, skipped, passed, total)
 	t.Log(summary)
-	fmt.Fprint(os.Stdout, summary)
+	fmt.Fprintln(os.Stdout, summary)
 
-	missing := total - (passed + failed + skipped)
+	missing = total - (passed + failed + skipped)
 	if missing != 0 {
 		for t := range pf.testNamesPassed {
 			delete(pf.testNames, t)
@@ -142,10 +203,12 @@ func (pf *ParallelTestFixture) PrintSummary(t *testing.T) {
 		}
 		t.Fatalf("Some tests did not report status: %s", strings.Join(missingTests, ", "))
 	}
+	return total, passed, skipped, failed, missing
 }
 
 func (pf *ParallelTestFixture) NewIsolatedFixture(t *testing.T, fcfg fixtureConfig) *TestFixture {
 	t.Helper()
 	pf.recordTestStarted(t)
-	return newTestFixture(t, pf, pf.NextAddr, fcfg)
+	envDesc := getEnvDesc()
+	return newTestFixture(t, envDesc, pf, pf.NextAddr, fcfg)
 }